refactor: extract sync service
